requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
Flask==2.3.3
gunicorn==21.2.0
